backend/requirements.txt
fastapi
uvicorn
requests
pika
psycopg2-binary
redis
python-dotenv
sqlalchemy
alembic
pytest
httpx
